ray generation and hit handling

+++ RayTracer/Matrix.h
@@ -12,7 +12,7 @@ class Matrix {
 
 	int row = 0;
 	int col = 0;
-	vector<vector<double>> data;
+	vector<vector<float>> data;
 
 public:
 
@@ -26,7 +26,7 @@ public:
 				{0, 0, 0, 1}};
 	}
 
-	Matrix(vector<vector<double>> data) {
+	Matrix(vector<vector<float>> data) {
 
 		row = data.size();
 		col = data[0].size();
@@ -41,7 +41,7 @@ public:
 		row = _row;
 		col = _col;
 		for (int i = 0; i < row; i++) {
-			vector<double> temp;
+			vector<float> temp;
 			for (int j = 0; j < col; j++) {
 				temp.push_back(0);
 			}
@@ -90,15 +90,15 @@ public:
 
 	Point operator*(Point const& other) const {  //matrix * tuple is tuple
 
-		vector<double> temp = other.getAll();
-		vector<double> final;
+		vector<float> temp = other.getAll();
+		vector<float> final;
 
 		if (row != 4) {
 			cout << "Column of first matrix should be equal to row of second matrix" << endl;
 		}
 		else {
 			for (int i = 0; i < row; i++) {
-				double num = 0;
+				float num = 0;
 				for (int j = 0; j < col; j++) {
 					num += data[i][j] * temp[j];
 				}
@@ -112,15 +112,15 @@ public:
 
 	Vector operator*(Vector const& other) const {  //matrix * tuple is tuple //applying transformations
 
-		vector<double> temp = other.getAll();
-		vector<double> final;
+		vector<float> temp = other.getAll();
+		vector<float> final;
 
 		if (row != 4) {
 			cout << "Column of first matrix should be equal to row of second matrix" << endl;
 		}
 		else {
 			for (int i = 0; i < row; i++) {
-				double num = 0;
+				float num = 0;
 				for (int j = 0; j < col; j++) {
 					num += data[i][j] * temp[j];
 				}
@@ -200,8 +200,8 @@ public:
 
 	/* Recursive function for finding determinant of matrix.
 	   n is current dimension of A[][]. */
-	double determinant(vector<vector<double>> A, int n) {
-		double D = 0; // Initialize result 
+	float determinant(vector<vector<float>> A, int n) {
+		float D = 0; // Initialize result 
 
 		//  Base case : if matrix contains single element 
 		if (n == 1) {
@@ -210,7 +210,7 @@ public:
 
 
 		int sign = 1;  // To store sign multiplier 
-		vector<vector<double>> temp = nullMat; // To store cofactors 
+		vector<vector<float>> temp = nullMat; // To store cofactors 
 
 		 // Iterate for each element of first row 
 		for (int f = 0; f < n; f++) {
@@ -225,12 +225,12 @@ public:
 		return D;
 	}
 
-	vector<vector<double>> nullMat = { {0, 0, 0, 0},
+	vector<vector<float>> nullMat = { {0, 0, 0, 0},
 									   {0, 0, 0, 0},
 									   {0, 0, 0, 0},
 									   {0, 0, 0, 0}};
 
-	void getCofactor(vector<vector<double>> A, vector<vector<double>>& temp, int p, int q, int n) {
+	void getCofactor(vector<vector<float>> A, vector<vector<float>>& temp, int p, int q, int n) {
 		int i = 0, j = 0;
 
 		// Looping for each element of the matrix 
@@ -257,10 +257,10 @@ public:
 	}
 
 	// Function to get adjoint of A[N][N] in adj[N][N]. 
-	void adjoint(vector<vector<double>> A, vector<vector<double>>& adj) {
+	void adjoint(vector<vector<float>> A, vector<vector<float>>& adj) {
 		// temp is used to store cofactors of A[][] 
 		int sign = 1;
-		vector<vector<double>> temp = nullMat;
+		vector<vector<float>> temp = nullMat;
 
 		for (int i = 0; i < 4; i++)
 		{
@@ -283,9 +283,9 @@ public:
 	// Function to calculate and store inverse, returns false if 
 	// matrix is singular 
 	Matrix inverse() {
-		vector<vector<double>> inverse = nullMat;
+		vector<vector<float>> inverse = nullMat;
 
-		vector<vector<double>> identity = { {1, 0, 0, 0},
+		vector<vector<float>> identity = { {1, 0, 0, 0},
 											{ 0, 1, 0, 0 },
 											{ 0, 0, 1, 0 },
 											{ 0, 0, 0, 1 } };
@@ -297,7 +297,7 @@ public:
 
 
 		// Find determinant of A[][] 
-		double det = determinant(data, 4);
+		float det = determinant(data, 4);
 		if (det == 0) {
 			cout << "Singular matrix, can't find its inverse" << endl;
 			//return false;
@@ -305,13 +305,13 @@ public:
 
 		// Find adjoint 
 
-		vector<vector<double>> adj = nullMat;
+		vector<vector<float>> adj = nullMat;
 		adjoint(data, adj);
 
 		// Find Inverse using formula "inverse(A) = adj(A)/det(A)" 
 		for (int i = 0; i < 4; i++) {
 			for (int j = 0; j < 4; j++) {
-				inverse[i][j] = adj[i][j] / double(det);
+				inverse[i][j] = adj[i][j] / float(det);
 			}
 		}
 
@@ -327,7 +327,7 @@ public:
 	//Vectors are translation independent (origin independent) so there is no point in translating
 	void translate(Point p) {
 		
-		Matrix translationMatrix = vector<vector<double>>{ {1, 0, 0, p.x},
+		Matrix translationMatrix = vector<vector<float>>{ {1, 0, 0, p.x},
 														   {0, 1, 0, p.y},
 														   {0, 0, 1, p.z},
 														   {0, 0, 0, 1}};
@@ -343,9 +343,9 @@ public:
 	
 	//Scaling
 	//TODO: Multiply scaling * vector
-	void scale(double x, double y, double z) {
+	void scale(float x, float y, float z) {
 
-		Matrix scalingMatrix = vector<vector<double>> {{x, 0, 0, 0},
+		Matrix scalingMatrix = vector<vector<float>> {{x, 0, 0, 0},
 													   {0, y, 0, 0},
 													   {0, 0, z, 0},
 													   {0, 0, 0, 1}};
@@ -359,14 +359,14 @@ public:
 	}
 
 	//TODO: Rotation Points
-	Matrix rotationMatrix(double x, double y, double z) {
+	Matrix rotationMatrix(float x, float y, float z) {
 	
 		Matrix X = Matrix().rotationXMatrix(x);
 		Matrix Y = Matrix().rotationYMatrix(y);
 		Matrix Z = Matrix().rotationZMatrix(z);
 		return X * Y * Z;
 	}
-	void rotate(double x, double y, double z) {
+	void rotate(float x, float y, float z) {
 
 		Matrix temp = rotationMatrix(x, y, z);
 		Matrix temp2 = data;
@@ -374,16 +374,16 @@ public:
 	}
 
 	//RotationX
-	Matrix rotationXMatrix(double radians) {
-		double const c = std::cos(radians);
-		double const s = std::sin(radians);
-		vector<vector<double>> temp = {{1, 0, 0, 0},
+	Matrix rotationXMatrix(float radians) {
+		float const c = std::cos(radians);
+		float const s = std::sin(radians);
+		vector<vector<float>> temp = {{1, 0, 0, 0},
 									  {0,  c , -s , 0},
 									  {0,  s ,  c , 0},
 									  {0, 0, 0, 1}};
 		return temp;
 	}
-	void rotateX(double radians) {
+	void rotateX(float radians) {
 
 		Matrix temp = rotationXMatrix(radians);
 		Matrix temp2 = data;
@@ -391,16 +391,16 @@ public:
 	}
 
 	//RotationY
-	Matrix rotationYMatrix(double radians) {
-		double const c = std::cos(radians);
-		double const s = std::sin(radians);
-		vector<vector<double>> temp = {{c, 0, s, 0},
+	Matrix rotationYMatrix(float radians) {
+		float const c = std::cos(radians);
+		float const s = std::sin(radians);
+		vector<vector<float>> temp = {{c, 0, s, 0},
 									  {0, 1, 0, 0},
 									  {-s, 0,  c, 0},
 									  {0, 0, 0, 1}};
 		return temp;
 	}
-	void rotateY(double radians) {
+	void rotateY(float radians) {
 
 		Matrix temp = rotationYMatrix(radians);
 		Matrix temp2 = data;
@@ -408,10 +408,10 @@ public:
 	}
 
 	//RotationZ
-	Matrix rotationZMatrix(double radians) {
-		double const c = std::cos(radians);
-		double const s = std::sin(radians);
-		vector<vector<double>> temp = {{c, -s, 0, 0},
+	Matrix rotationZMatrix(float radians) {
+		float const c = std::cos(radians);
+		float const s = std::sin(radians);
+		vector<vector<float>> temp = {{c, -s, 0, 0},
 									  {s,  c, 0, 0},
 									  {0, 0, 1, 0},
 									  {0, 0, 0, 1}};
@@ -422,7 +422,7 @@ public:
 		// | 1 
 		return temp;
 	}
-	void rotateZ(double radians) {
+	void rotateZ(float radians) {
 
 		Matrix temp = rotationZMatrix(radians);
 		Matrix temp2 = data;
@@ -430,9 +430,9 @@ public:
 	}
 
 	//Shearing
-	void shear(double xy, double xz, double yx, double yz, double zx, double zy) {
+	void shear(float xy, float xz, float yx, float yz, float zx, float zy) {
 
-		Matrix shearingMatrix = vector<vector<double>> {{1, xy, xz, 0},
+		Matrix shearingMatrix = vector<vector<float>> {{1, xy, xz, 0},
 														{yx, 1,  yz, 0},
 														{zx, zy, 1, 0},
 														{0, 0, 0, 1}};
